edit: exercise/collatz-conjecture - size_t usage

+++ collatz-conjecture/collatz_conjecture.c
@@ -1,11 +1,12 @@
 #include "collatz_conjecture.h"
+#include <stddef.h>
 
 int steps(int start)
 {
   if (start < 1)
     return ERROR_VALUE;
 
-  int count = 0;
+  size_t count = 0;
   for (; start > 1; count++)
     start = (start % 2) ? start * 3 + 1 : start / 2;
